+++ src/credential.cpp
@@ -1,4 +1,5 @@
 #include "credential.h"
+#include "tls_syntax.h"
 
 #define DUMMY_SIG_SCHEME SignatureScheme::P256_SHA256
 
@@ -26,6 +27,29 @@ operator>>(tls::istream& in, CredentialType& type)
 ///
 /// BasicCredential
 ///
+
+// struct {
+//     opaque identity<0..2^16-1>;
+//     SignatureScheme algorithm;
+//     SignaturePublicKey public_key;
+// } BasicCredential;
+class BasicCredential : public AbstractCredential
+{
+public:
+  BasicCredential();
+  BasicCredential(const bytes& identity, const SignaturePublicKey& public_key);
+
+  virtual std::unique_ptr<AbstractCredential> dup() const;
+  virtual bytes identity() const;
+  virtual SignaturePublicKey public_key() const;
+  virtual void read(tls::istream& in);
+  virtual void write(tls::ostream& out) const;
+  virtual bool equal(const AbstractCredential* other) const;
+
+private:
+  tls::opaque<2> _identity;
+  SignaturePublicKey _public_key;
+};
 
 BasicCredential::BasicCredential()
   : _public_key(DUMMY_SIG_SCHEME)
